Periodic Syncing - Mitigated Module Spasm
When Modules approach the reset position, the massive change in current position screws with the PID since the reference set by it is relative to the current position. In this commit, I have delayed the syncing for a wheel that approaches its reset position. When the module is near its reset position, it will not sync, allowing it to go beyond the reset boundaries briefly to mitigate PID's spasm. I have concluded that for a periodically called method at 20ms, it is impossible to remove PID's spasm as it is a result of lag when PID's reference is targeting a specific location when the module has already reset. If each tick were much faster than 20ms, the reset should not affect PID because PID is always referenced to the current position no matter where it is as fast as possible.

When modules leave their reset "zone," they will jerk a little -  sometimes even still spasm but quickly calm down.